txn: extend fd_txn_parse to allow trailing bytes
Adds a payload_size_opt output parameter to fd_txn_parse that permits
providing buffers with trailing bytes after the txn. This is usually the
case when the size of the txn is not known beforehand. For example, the
wire representation of a block relies on the self-describing properties
of the txn serialization to concatenate txns without length prefixes.