ffi: add idiomatic wrapper (not safe)
TBD. Draft. Not safe. Not tested. Not documented. Not ready for use.

Nicer to use wrapper around the raw FFI bindings. It's not really safe
and difficult to think about how you could make it safe, but its at
least more usable for now. This will hopefully be used by the Solana
Labs bank thread to run as a tile.