xdp: also listen on loopback by default
The Linux kernel does some short circuiting optimizations
when sending packets to an IP address that's owned by the
same host. The optimization is basically to route them over
to the loopback interface directly, bypassing the network
hardware.

This redirection to the loopback interface happens before
XDP programs are executed, so local traffic destined for
our listen addresses will not get ingested correctly.

There are two reasons we send traffic locally,

* For testing and development.
* The Solana Labs code sends local traffic to itself to
  as part of routine operation (eg, when it's the leader
  it sends votes to its own TPU socket).

So for now we need to also bind to loopback. This is a
small performance hit for other traffic, but we only
redirect packets destined for our target IP and port so
it will not otherwise interfere.

Hopefully we can remove this for production use cases
soon once we change how the validator communicates
with itself.